# openapi-server/target/generated-sources/openapi/src/main/java/com/example/openapi/generated/api/GreetingsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.openapi.generated.api;

import com.example.openapi.generated.model.GreetingRequest;
import com.example.openapi.generated.model.GreetingResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "Greetings", description = "Operações relacionadas a cumprimento")
public interface GreetingsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /greetings : Persiste uma saudação customizada
     *
     * @param greetingRequest  (required)
     * @return Saudação criada (status code 201)
     *         or Erro de validação (status code 400)
     */
    @Operation(
        operationId = "createGreeting",
        summary = "Persiste uma saudação customizada",
        tags = { "Greetings" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Saudação criada", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GreetingResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Erro de validação")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/greetings",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<GreetingResponse> createGreeting(
        @Parameter(name = "GreetingRequest", description = "", required = true) @Valid @RequestBody GreetingRequest greetingRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"language\" : \"language\", \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /greetings/{name} : Recupera uma saudação personalizada
     *
     * @param name Nome da pessoa (required)
     * @param title Título opcional (Dr., Ms., etc) (optional)
     * @return Saudação calculada (status code 200)
     *         or Não encontrado (status code 404)
     */
    @Operation(
        operationId = "getGreeting",
        summary = "Recupera uma saudação personalizada",
        tags = { "Greetings" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Saudação calculada", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GreetingResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Não encontrado")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/greetings/{name}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<GreetingResponse> getGreeting(
        @Parameter(name = "name", description = "Nome da pessoa", required = true, in = ParameterIn.PATH) @PathVariable("name") String name,
        @Parameter(name = "title", description = "Título opcional (Dr., Ms., etc)", in = ParameterIn.QUERY) @Valid @RequestParam(value = "title", required = false) String title
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"language\" : \"language\", \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}